Introduced Geometry templates: - Added Cube Geometry. - Added Mesh Geometry. - Added 2D Batching support to BatchManager.
Added entry point for Importer/Exported classes that import/export model files (currently only in .gltf)

Performance Optimization:
- Replaced virtual functions event system design into more "static bound" way in order to increase function call speed during GLFW callbacks.

Convention refactoring:
- removed prefixes "x" and "c" from classes/structs as they make code reading a little more complicated.
- now all our files are stored in directory "resources" (any files used for loading). So if you wanna read/write models or images or anything like that, you use "resources/" filepath. Also resources are used for "import/export" process only during "development cycle of game". So during actual "game runtime" we will use another directory path called "assets/" where we will store game resources in the safe and fast custom binary format.

Testing:
- Added test_config.json and TestConfigReader parser in order to store "testing settings" during runtime testing.
- Added AnimateLight flag into test configs in order to simulate light movement and color changes in update loop.

+++ engine/cpp/geometry/geometry.cpp
@@ -4,7 +4,7 @@ namespace xpe {
 
     namespace math {
 
-        void IndexArray::Init(int count) {
+        void IndexArray::Init(usize count) {
             Data.resize(count);
         }
